requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
altair=5.1.2=pypi_0
ansi2html=1.8.0=pypi_0
asttokens=2.4.0=pyhd8ed1ab_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
beautifulsoup4=4.12.2=pypi_0
blas=1.0=openblas
blinker=1.7.0=pypi_0
bottleneck=1.3.5=py310ha9d4c09_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=h5eee18b_0
ca-certificates=2023.7.22=hbcca054_0
cachetools=5.3.1=pypi_0
cairo=1.16.0=hb05425b_5
certifi=2023.7.22=pypi_0
charset-normalizer=3.3.0=pypi_0
click=8.1.7=pypi_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.0.5=py310hdb19cb5_0
customtkinter=5.2.1=pypi_0
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
darkdetect=0.8.0=pypi_0
dash=2.14.1=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
dbus=1.13.18=hb2f20db_0
debugpy=1.6.7=py310h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pypi_0
eigen=3.3.7=hd09550d_1
et-xmlfile=1.1.0=pypi_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
expat=2.5.0=h6a678d5_0
ffmpeg=4.2.2=h20bf706_0
filelock=3.12.4=pypi_0
flask=3.0.0=pypi_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.10.0=pypi_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.11=pypi_0
gitpython=3.1.40=pypi_0
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google=3.0.0=pypi_0
google-api-core=2.12.0=pypi_0
google-auth=2.23.3=pypi_0
google-cloud-vision=3.4.5=pypi_0
googleapis-common-protos=1.61.0=pypi_0
graphite2=1.3.14=h295c915_1
grpcio=1.59.0=pypi_0
grpcio-status=1.59.0=pypi_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
harfbuzz=4.3.0=hf52aaf7_1
hdf5=1.12.1=h2b7332f_3
huggingface-hub=0.17.3=pypi_0
icu=58.2=he6710b0_3
idna=3.4=pypi_0
imageio=2.34.0=pypi_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
ipykernel=6.25.2=pyh2140261_0
ipython=8.16.1=pyh0d859eb_0
itsdangerous=2.1.2=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=py310h06a4308_0
jpeg=9e=h5eee18b_1
jsonschema=4.20.0=pypi_0
jsonschema-specifications=2023.11.1=pypi_0
jupyter_client=8.4.0=pyhd8ed1ab_0
jupyter_core=5.4.0=py310hff52083_0
kiwisolver=1.4.4=py310h6a678d5_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lazy-loader=0.3=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
levenshtein=0.23.0=pypi_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcups=2.4.2=h2d74bed_1
libcurl=8.4.0=h251f7ec_0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=hdbd6064_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libidn2=2.3.4=h5eee18b_0
libllvm14=14.0.6=hdb19cb5_3
libnghttp2=1.57.0=h2d74bed_0
libopenblas=0.3.21=h043d6bf_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.39=h5eee18b_0
libpq=12.15=hdbd6064_1
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h36c2ea0_1
libssh2=1.10.0=hdbd6064_2
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libvpx=1.7.0=h439df22_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hcbfbd50_0
libxslt=1.1.37=h2085143_0
lz4-c=1.9.4=h6a678d5_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.2=py310h06a4308_0
matplotlib-base=3.7.2=py310h1128e8f_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
munkres=1.1.4=py_0
mysql=5.7.24=h721c034_2
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=3.2=pypi_0
nspr=4.35=h6a678d5_0
nss=3.89.1=h6a678d5_0
numexpr=2.8.7=py310h286c3b5_0
numpy=1.26.0=py310heeff2f4_0
numpy-base=1.26.0=py310h8a23956_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.18.1=pypi_0
nvidia-nvjitlink-cu12=12.3.52=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
odfpy=1.4.1=pypi_0
opencv=4.6.0=py310ha16400d_5
opencv-python-headless=4.8.1.78=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openpyxl=3.1.2=pypi_0
openssl=3.0.11=h7f8727e_2
packaging=23.1=py310h06a4308_0
pandas=2.1.1=py310h1128e8f_0
parso=0.8.3=pyhd8ed1ab_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.0.1=pypi_0
pip=23.3=py310h06a4308_0
pixman=0.40.0=h7f8727e_1
platformdirs=3.11.0=pyhd8ed1ab_0
plotly=5.18.0=pypi_0
ply=3.11=py310h06a4308_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
proto-plus=1.22.3=pypi_0
protobuf=3.20.0=pypi_0
psutil=5.9.0=py310h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=14.0.1=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pydeck=0.8.1b0=pypi_0
pygments=2.16.1=pyhd8ed1ab_0
pyparsing=3.0.9=py310h06a4308_0
pyqt=5.15.7=py310h6a678d5_1
pyqt5-sip=12.11.0=pypi_0
pytesseract=0.3.10=pypi_0
python=3.10.13=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.10=2_cp310
pytz=2023.3.post1=py310h06a4308_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.0=py310h6a678d5_0
qt-main=5.15.2=h7358343_9
qt-webengine=5.15.9=h9ab4d14_7
qtwebkit=5.212=h3fafdc1_5
rapidfuzz=3.6.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.31.0=pypi_0
regex=2023.10.3=pypi_0
requests=2.31.0=pypi_0
retrying=1.3.4=pypi_0
rich=13.7.0=pypi_0
rpds-py=0.13.0=pypi_0
rsa=4.9=pypi_0
sacremoses=0.1.1=pypi_0
safetensors=0.4.0=pypi_0
scikit-image=0.22.0=pypi_0
scikit-learn=1.3.0=py310h1128e8f_0
scipy=1.11.3=py310heeff2f4_0
seaborn=0.12.2=py310h06a4308_0
sentencepiece=0.1.96=pypi_0
setuptools=68.0.0=py310h06a4308_0
sip=6.6.2=py310h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.1=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
streamlit=1.30.0=pypi_0
sympy=1.12=pypi_0
tenacity=8.2.3=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2024.2.12=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.14.1=pypi_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.12.0=pypi_0
torch=2.1.0=pypi_0
tornado=6.3.3=py310h5eee18b_0
tqdm=4.66.1=pypi_0
traitlets=5.11.2=pyhd8ed1ab_0
transformers=4.16.2=pypi_0
triton=2.1.0=pypi_0
typer=0.4.1=pypi_0
typing-extensions=4.8.0=hd8ed1ab_0
typing_extensions=4.8.0=pyha770c72_0
tzdata=2023c=h04d1e81_0
tzlocal=5.2=pypi_0
urllib3=2.0.7=pypi_0
validators=0.22.0=pypi_0
watchdog=3.0.0=pypi_0
wcwidth=0.2.8=pyhd8ed1ab_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py310h06a4308_0
x264=1!157.20191217=h7b6447c_0
xz=5.4.2=h5eee18b_0
zeromq=4.3.4=h2531618_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
